Create Support & Resistance_TrendBars
Support & Resistance levels are most important for a trader and Trend Bars makes it super cool. Green bars shows the all well and Red bars you know what it means.

The script will draw the support and resistance blocks with prices on your grave and also HH, HL, LH, LL levels to keep you tuned on the road.

Diagonal Support and Resistance Trendlines with Trendy bars have added to show you the trend which gives you more confidence.

There is zigzag line which has been colored white to reduce mess on the graph so if you want to see the zigzag , just change the color from white to any you like.

You can use it on any timeframe but 1 Hour, 2 Hours, 4 Hours and Daily are recommended.